remove fragmenter from meta
