+++ proto/meta.proto
@@ -150,17 +150,12 @@ message TableFragments {
     ActorState state = 2;
   }
   message Fragment {
-    enum FragmentType {
-      SOURCE = 0;
-      SINK = 1;
-      OTHERS = 2;
-    }
     enum FragmentDistributionType {
       SINGLE = 0;
       HASH = 1;
     }
     uint32 fragment_id = 1;
-    FragmentType fragment_type = 2;
+    stream_plan.FragmentType fragment_type = 2;
     FragmentDistributionType distribution_type = 3;
     repeated stream_plan.StreamActor actors = 4;
   }
@@ -180,7 +175,7 @@ message CreateMaterializedViewRequest {
   uint32 node_id = 1;
   // TODO: remove this, do catalog init in Meta and return in response.
   plan_common.TableRefId table_ref_id = 2;
-  stream_plan.StreamNode stream_node = 3;
+  stream_plan.StreamFragmentGraph fragment_graph = 3;
 }
 
 // will be deprecated and replaced by DdlService
