+++ proto/stream_plan.proto
@@ -278,3 +278,42 @@ message StreamActor {
   // Placement rule for actor, need to stay on the same node as upstream.
   bool same_worker_node_as_upstream = 7;
 }
+
+enum FragmentType {
+  SOURCE = 0;
+  SINK = 1;
+  OTHERS = 2;
+}
+
+message StreamFragmentGraph {
+  message StreamFragment {
+    uint32 fragment_id = 1;
+    // root stream node in this fragment.
+    StreamNode node = 2;
+    FragmentType fragment_type = 3;
+    // mark whether this fragment should only have one actor.
+    bool is_singleton = 4;
+    // Number of table ids (stateful states) for this fragment.
+    uint32 table_ids_cnt = 5;
+  }
+
+  message StreamFragmentEdge {
+    // Dispatch strategy for the fragment.
+    DispatchStrategy dispatch_strategy = 1;
+    // Whether the two linked nodes should be placed on the same worker node
+    bool same_worker_node = 2;
+    // A unique identifer of this edge. Generally it should be exchange node's operator id. When
+    // rewriting fragments into delta joins or when inserting 1-to-1 exchange, there will be
+    // virtual links generated.
+    uint64 link_id = 3;
+    uint32 upstream_id = 4;
+    uint32 downstream_id = 5;
+  }
+  // all the fragments in the graph.
+  map<uint32, StreamFragment> fragments = 1;
+  // edges between fragments.
+  repeated StreamFragmentEdge edges = 2;
+
+  repeated uint32 dependent_table_ids=3;
+  uint32 table_ids_cnt=4;
+}
