+++ src/meta/Cargo.toml
@@ -63,6 +63,7 @@ workspace-hack = { version = "0.1", path = "../workspace-hack" }
 [dev-dependencies]
 assert_matches = "1"
 rand = "0.8"
+risingwave_frontend = { path = "../frontend" }
 tempfile = "3"
 
 [features]
